+++ Green-Nacho-Engine/Model.cpp
@@ -38,13 +38,13 @@ namespace gn
 
 	void Model::processNode(aiNode* node, const aiScene* scene, bool hasTextures)
 	{
-		for (int i = 0; i < node->mNumMeshes; i++)
+		for (int i = 0; i < (int)node->mNumMeshes; i++)
 		{
 			aiMesh* mesh = scene->mMeshes[node->mMeshes[i]];
 			_modelMeshes.push_back(generateMesh(mesh, scene, hasTextures));
 		}
 
-		for (int i = 0; i < node->mNumChildren; i++)
+		for (int i = 0; i < (int)node->mNumChildren; i++)
 			processNode(node->mChildren[i], scene, hasTextures);
 	}	
 	
@@ -53,7 +53,7 @@ namespace gn
 		std::vector<ModelMeshVertex> vertices;
 		std::vector<unsigned int> indexes;
 
-		for (int i = 0; i < mesh->mNumVertices; i++)
+		for (int i = 0; i < (int)mesh->mNumVertices; i++)
 		{
 			ModelMeshVertex vertex;
 			aiVector3D aiMeshVertex = mesh->mVertices[i];
@@ -70,11 +70,11 @@ namespace gn
 			vertices.push_back(vertex);
 		}
 
-		for (int i = 0; i < mesh->mNumFaces; i++)
+		for (int i = 0; i < (int)mesh->mNumFaces; i++)
 		{
 			aiFace face = mesh->mFaces[i];
 
-			for (int j = 0; j < face.mNumIndices; j++)
+			for (int j = 0; j < (int)face.mNumIndices; j++)
 				indexes.push_back(face.mIndices[j]);
 		}
 
@@ -101,7 +101,7 @@ namespace gn
 	{
 		std::vector<Texture*> textures;
 
-		for (int i = 0; i < material->GetTextureCount(type); i++)
+		for (int i = 0; i < (int)material->GetTextureCount(type); i++)
 		{
 			Texture* texture;
 			aiString string;
@@ -119,13 +119,13 @@ namespace gn
 
 	void Model::dispose()
 	{
-		for (int i = 0; i < _modelMeshes.size(); i++)
+		for (int i = 0; i < (int)_modelMeshes.size(); i++)
 			delete _modelMeshes[i];
 	}
 
 	void Model::draw() const
 	{
-		for (int i = 0; i < _modelMeshes.size(); i++)
+		for (int i = 0; i < (int)_modelMeshes.size(); i++)
 			_modelMeshes[i]->draw();
 	}
 }
